editor, ui ilerlemesi
-Basit bir harita yapılabiliyor ve kaydedilebiliyor.
-Panelin etrafına daha efektik bir şekilde outline çizebiliyorum ancak çizgi kalınlığını bir türlü ayarlayamadım. Aklımda bir şeyler var ama şimdilik ileriye atıyorum.
-Rapidxml, okurken değil ama yazarken çok uğraştırdı. Herkeste olan bir bug bende de olduğu için kaynak dosyada ufak bir değişiklik ile çözdüm. 
-Butonlar render edilirken, artık bir renk ile beraber render etmek zorunda değil. (Eskiden mecburi bir renk vardı ve görüntüyü bozabiliyordu. Özellikle sprite'ı sabit kalmasını istediğimiz bir butonda oluyordu.)
-Panele isMouseHover() eklendi.
-ScreenToWorld(), CellToPosition(), PositionToCell() eklendi. WorldToScreen(), sadece göz boyamak için duruyor. Düzenlenecek. 
-TextRenderer'daki meşhur casting uyarısını çözdüm. Şimdiye kadar üşeniyordum sadece.

+++ src/Renderers/TextRenderer.cpp
@@ -96,7 +96,7 @@ void TextRenderer::Load(std::string font, GLuint fontSize)
 			texture,
 			glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 			glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-			face->glyph->advance.x};
+			static_cast<GLuint>(face->glyph->advance.x)};
 		Characters.insert(std::pair<GLchar, Character>(c, character));
 	}
 	glBindTexture(GL_TEXTURE_2D, 0);
